+++ FastAnimations/ModEntry.cs
@@ -93,6 +93,8 @@ namespace Pathoschild.Stardew.FastAnimations
         {
             if (config.BreakGeodeSpeed > 1)
                 yield return new BreakingGeodeHandler(config.BreakGeodeSpeed);
+            if (config.CasinoSlotsSpeed > 1)
+                yield return new CasinoSlotsHandler(config.CasinoSlotsSpeed, this.Helper.Reflection);
             if (config.EatAndDrinkSpeed > 1)
                 yield return new EatingHandler(this.Helper.Reflection, config.EatAndDrinkSpeed, config.DisableEatAndDrinkConfirmation);
             if (config.FishingSpeed > 1)