+++ FastAnimations/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## 1.4.1
 * Updated to SMAPI 2.4.
+* Added support for casino slots.
 * Improved fast tree falling performance.
 
 ## 1.4